Fix for setting and not accidentally resetting points when creating new contact that exists in Mautic via API
